Instruction execution RAF sumcheck, prefix-suffix abstraction (#738)
* prefix suffix for identity poly


* raf integration - errors somewhere rn


* raf integration working e2e


* downcast indices for all instructions


* split batched poly into two opernads polys


* optimize prefix calculation


* paralelize and add rwlock


* clippy


* perf: replace Mutex<Option<Vec<F>>> with OnceCell<Vec<F>> in CachedPolynomial

- OnceCell provides better performance for write-once read-many pattern
- Eliminates lock contention for read operations after initial computation
- Simplifies the caching logic with get_or_init pattern
- Use vec\![OnceCell::new(); capacity] for faster initialization
- Remove unnecessary tracing span for allocation

* rename for simplicity


* once cell take


* use arrays instead of vecs


* is even


* rm set_bound_value fn


* clean up sumcheck prover message fn


---------

+++ jolt-core/src/jolt/instruction/virtual_assert_halfword_alignment.rs
@@ -42,11 +42,11 @@ impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RISCVCycle<VirtualAssert
             #[cfg(test)]
             8 => (
                 self.register_state.rs1 as u8 as u64,
-                self.instruction.operands.imm,
+                self.instruction.operands.imm as u8 as i64,
             ),
             32 => (
                 self.register_state.rs1 as u32 as u64,
-                self.instruction.operands.imm,
+                self.instruction.operands.imm as u32 as i64,
             ),
             64 => (self.register_state.rs1, self.instruction.operands.imm),
             _ => panic!("{WORD_SIZE}-bit word size is unsupported"),
